requirements.txt
tensorflow==2.15.0
numpy>=1.24.3
pandas>=2.0.0
scikit-learn>=1.3.0
opencv-python>=4.8.0
pillow>=10.0.0
matplotlib>=3.7.1
mlflow>=2.9.0
pytest>=7.4.0
black>=23.7.0
isort>=5.12.0
flake8>=6.1.0
dvc>=3.30.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-dotenv>=1.0.0
wandb>=0.15.12
hydra-core>=1.3.2
albumentations>=1.3.1
pytest-cov>=4.1.0
pre-commit>=3.3.3 